# PERF cache find_stack_level