--- a/pandas/_testing/asserters.py
+++ b/pandas/_testing/asserters.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import inspect
 from typing import (
     Literal,
     cast,
@@ -112,7 +113,7 @@ def assert_almost_equal(
             "is deprecated and will be removed in a future version. "
             "You can stop passing 'check_less_precise' to silence this warning.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
         rtol = atol = _get_tol_from_less_precise(check_less_precise)
 
@@ -339,7 +340,7 @@ def assert_index_equal(
             "is deprecated and will be removed in a future version. "
             "You can stop passing 'check_less_precise' to silence this warning.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
         rtol = atol = _get_tol_from_less_precise(check_less_precise)
 
@@ -815,7 +816,7 @@ def assert_extension_array_equal(
             "is deprecated and will be removed in a future version. "
             "You can stop passing 'check_less_precise' to silence this warning.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
         rtol = atol = _get_tol_from_less_precise(check_less_precise)
 
@@ -970,7 +971,7 @@ def assert_series_equal(
             "is deprecated and will be removed in a future version. "
             "You can stop passing 'check_less_precise' to silence this warning.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
         rtol = atol = _get_tol_from_less_precise(check_less_precise)
 
@@ -1263,7 +1264,7 @@ def assert_frame_equal(
             "is deprecated and will be removed in a future version. "
             "You can stop passing 'check_less_precise' to silence this warning.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
         rtol = atol = _get_tol_from_less_precise(check_less_precise)