--- a/pandas/core/accessor.py
+++ b/pandas/core/accessor.py
@@ -6,6 +6,7 @@ that can be mixed into or pinned onto other pandas classes.
 """
 from __future__ import annotations
 
+import inspect
 import warnings
 
 from pandas.util._decorators import doc
@@ -268,7 +269,7 @@ def _register_accessor(name, cls):
                 f"{repr(name)} for type {repr(cls)} is overriding a preexisting "
                 f"attribute with the same name.",
                 UserWarning,
-                stacklevel=find_stack_level(),
+                stacklevel=find_stack_level(inspect.currentframe()),
             )
         setattr(cls, name, CachedAccessor(name, accessor))
         cls._accessors.add(name)