--- a/pandas/core/arrays/arrow/_arrow_utils.py
+++ b/pandas/core/arrays/arrow/_arrow_utils.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import inspect
 import json
 import warnings
 
@@ -22,7 +23,9 @@ def fallback_performancewarning(version: str | None = None) -> None:
     msg = "Falling back on a non-pyarrow code path which may decrease performance."
     if version is not None:
         msg += f" Upgrade to pyarrow >={version} to possibly suppress this warning."
-    warnings.warn(msg, PerformanceWarning, stacklevel=find_stack_level())
+    warnings.warn(
+        msg, PerformanceWarning, stacklevel=find_stack_level(inspect.currentframe())
+    )
 
 
 def pyarrow_array_to_numpy_and_mask(
@@ -133,7 +136,7 @@ class ArrowIntervalType(pyarrow.ExtensionType):
         warnings.warn(
             "Attribute `closed` is deprecated in favor of `inclusive`.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
         return self._inclusive