--- a/pandas/core/computation/eval.py
+++ b/pandas/core/computation/eval.py
@@ -3,6 +3,7 @@ Top level ``eval`` module.
 """
 from __future__ import annotations
 
+import inspect
 import tokenize
 from typing import TYPE_CHECKING
 import warnings
@@ -311,7 +312,7 @@ def eval(
                 "will be removed in a future version."
             ),
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
 
     exprs: list[str | BinOp]