--- a/pandas/core/dtypes/concat.py
+++ b/pandas/core/dtypes/concat.py
@@ -3,6 +3,7 @@ Utility functions related to concat.
 """
 from __future__ import annotations
 
+import inspect
 from typing import (
     TYPE_CHECKING,
     cast,
@@ -152,7 +153,7 @@ def concat_compat(to_concat, axis: int = 0, ea_compat_axis: bool = False):
             "(instead of coercing bools to numeric values). To retain the old "
             "behavior, explicitly cast bool-dtype arrays to numeric dtype.",
             FutureWarning,
-            stacklevel=find_stack_level(),
+            stacklevel=find_stack_level(inspect.currentframe()),
         )
     return result