… pandas/core/dtypes/astype.py
… pandas/core/dtypes/cast.py
… pandas/core/dtypes/common.py
… pandas/core/dtypes/dtypes.py
… pandas/core/dtypes/inference.py
… pandas/core/frame.py
… pandas/core/generic.py
… pandas/core/groupby/generic.py
… pandas/core/groupby/groupby.py
… pandas/core/groupby/grouper.py
… pandas/core/index.py
… pandas/core/indexers/utils.py
… pandas/core/indexes/accessors.py
… pandas/core/indexes/base.py
… pandas/core/indexes/category.py
… pandas/core/indexes/datetimelike.py
… pandas/core/indexes/datetimes.py
… pandas/core/indexes/interval.py
… pandas/core/indexes/multi.py
… pandas/core/indexes/numeric.py
… pandas/core/indexes/period.py
… pandas/core/indexes/range.py
… pandas/core/indexing.py
… pandas/core/internals/__init__.py
… pandas/core/internals/blocks.py
… pandas/core/internals/construction.py
… pandas/core/internals/managers.py
… pandas/core/ops/__init__.py
… pandas/core/resample.py
… pandas/core/reshape/concat.py