@MarcoGorelli 2022-10-14
Thanks for reporting, version 1.5.1 (probably coming out Monday) will fix this